+++ changelog.md
@@ -3,6 +3,9 @@
 
 > **Note:** The dates included here are the dates of the commits that were made to the repository. The actual release dates may be different.
 
+# 28 August 2025 (dev)
+* Add Helios Software fhirpath engine for R4B/R6
+
 # 20 August 2025 (dev)
 * refactor the engine selector to remove some of the old engine selector code that's no longer needed
 * move the debugger controls into the header nav bar from the tab header (moving the other buttons is bad user experience)